ci(tests): reorganize acceptance tests
Changelog: None
Ticket: None
